% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequoiaInput.R
\name{toSequoiaInput}
\alias{toSequoiaInput}
\title{Build Sequioa Input files for Tom's template script}
\usage{
toSequoiaInput(baselinePops, mixturePops, markerList, prefix = "")
}
\arguments{
\item{baselinePops}{Poplist of potential parents}

\item{mixturePops}{Poplist of offspring}

\item{markerList}{list of markers to use}

\item{prefix}{This is a string that will be used as the prefix for the two output files.
Default is to have no prefix.}
}
\value{
Writes files to the working directory that can be uploaded to the server and
  used to run single parentage with Sequoia using Tom's template script
}
\description{
This interfaces with IDFGEN objects to build files expected by Tom's template Sequoia script
  for single parent assignment (actual assignments, not simulations)
}
