% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequoiaInput.R
\name{toSequoiaInput_EFGLmh}
\alias{toSequoiaInput_EFGLmh}
\title{Build Sequioa Input files for Tom's template script}
\usage{
toSequoiaInput_EFGLmh(
  x,
  baselinePops,
  mixturePops,
  markerList = NULL,
  prefix = ""
)
}
\arguments{
\item{x}{EFGLdata object with potential parents and offspring}

\item{baselinePops}{names of pops with potential parents}

\item{mixturePops}{names of pops with offspring}

\item{markerList}{list of markers to use, if not provided the default is to use all}

\item{prefix}{This is a string that will be used as the prefix for the two output files.
Default is to have no prefix.}
}
\value{
Writes files to the working directory that can be uploaded to the server and
  used to run single parentage with Sequoia using Tom's template script
}
\description{
This interfaces with EFGLmh objects to build files expected by Tom's template Sequoia script
  for single parent assignment (actual assignments, not simulations)
}
