% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequoiaSimulation.R
\name{toSequoiaSimulation}
\alias{toSequoiaSimulation}
\title{Build Sequioa Input files for \code{sequoiaSim}}
\usage{
toSequoiaSimulation(baselinePops, markerList)
}
\arguments{
\item{baselinePops}{Poplist of potential parents}

\item{markerList}{list of markers to use}
}
\value{
Writes a file to the working directory that can be uploaded to the server and
  used to run \code{sequoiaSim}
}
\description{
This interfaces with IDFGEN objects to build the parent_data_file expected by \code{sequoiaSim}
}
