% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequoiaSimulation.R
\name{toSequoiaSimulation_EFGLmh}
\alias{toSequoiaSimulation_EFGLmh}
\title{Build Sequioa Input files for \code{sequoiaSim}}
\usage{
toSequoiaSimulation_EFGLmh(x, baselinePops, markerList = NULL, prefix = "")
}
\arguments{
\item{x}{EFGLdata object with potential parents and offspring}

\item{baselinePops}{names of pops with potential parents}

\item{markerList}{list of markers to use}

\item{prefix}{This is a string that will be used as the prefix for the output files.
Default is to have no prefix.}
}
\value{
Writes a file to the working directory that can be uploaded to the server and
  used to run \code{sequoiaSim}
}
\description{
This interfaces with EFGLmh objects to build the parent_data_file expected by \code{sequoiaSim}
}
